+++ src/main/errors.cpp
@@ -87,7 +87,7 @@ void R_CheckStack(void)
     long usage = R_CStackDir * (R_CStackStart - (unsigned long)&dummy);
 
     /* printf("usage %ld\n", usage); */
-    if(R_CStackLimit != -1 && usage > 0.95 * R_CStackLimit) {
+    if(R_CStackLimit != uintptr_t(-1) && usage > 0.95 * R_CStackLimit) {
 	/* We do need some stack space to process error recovery,
 	   so temporarily raise the limit.
 	 */
@@ -229,7 +229,7 @@ void warning(const char *format, ...)
     va_end(ap);
     p = buf + strlen(buf) - 1;
     if(strlen(buf) > 0 && *p == '\n') *p = '\0';
-    if(R_WarnLength < BUFSIZE - 20 && strlen(buf) == R_WarnLength)
+    if(R_WarnLength < BUFSIZE - 20 && int(strlen(buf)) == R_WarnLength)
 	strcat(buf, " [... truncated]");
     warningcall(R_NilValue, buf);
 }
@@ -289,7 +289,7 @@ static void vwarningcall_dflt(SEXP call, const char *format, va_list ap)
 
     if(w >= 2) { /* make it an error */
 	Rvsnprintf(buf, min(BUFSIZE, R_WarnLength), format, ap);
-	if(R_WarnLength < BUFSIZE - 20 && strlen(buf) == R_WarnLength)
+	if(R_WarnLength < BUFSIZE - 20 && int(strlen(buf)) == R_WarnLength)
 	    strcat(buf, " [... truncated]");
 	inWarning = 0; /* PR#1570 */
 	errorcall(call, _("(converted from warning) %s"), buf);
@@ -303,7 +303,7 @@ static void vwarningcall_dflt(SEXP call, const char *format, va_list ap)
 	else
 	    REprintf(_("Warning: "));
 	Rvsnprintf(buf, min(BUFSIZE, R_WarnLength+1), format, ap);
-	if(R_WarnLength < BUFSIZE - 20 && strlen(buf) == R_WarnLength)
+	if(R_WarnLength < BUFSIZE - 20 && int(strlen(buf)) == R_WarnLength)
 	    strcat(buf, " [... truncated]");
 	REprintf("%s\n", buf);
     }
@@ -314,7 +314,7 @@ static void vwarningcall_dflt(SEXP call, const char *format, va_list ap)
 	    return;
 	SET_VECTOR_ELT(R_Warnings, R_CollectWarnings, call);
 	Rvsnprintf(buf, min(BUFSIZE, R_WarnLength+1), format, ap);
-	if(R_WarnLength < BUFSIZE - 20 && strlen(buf) == R_WarnLength)
+	if(R_WarnLength < BUFSIZE - 20 && int(strlen(buf)) == R_WarnLength)
 	    strcat(buf, " [... truncated]");
 	names = CAR(ATTRIB(R_Warnings));
 	SET_STRING_ELT(names, R_CollectWarnings++, mkChar(buf));
